[FE] rename: Metor -> Mentor
